move media related components to media module
