+++ src/backoffice/media/media/components/index.ts
@@ -1,0 +1,1 @@
+import './input-media-picker/input-media-picker.element';
