+++ src/backoffice/media/media/components/input-media-picker/input-media-picker.element.ts
@@ -3,7 +3,7 @@ import { UUITextStyles } from '@umbraco-ui/uui-css/lib';
 import { customElement, property, state } from 'lit/decorators.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { FormControlMixin } from '@umbraco-ui/uui-base/lib/mixins';
-import { UmbMediaRepository } from '../../../media/media/repository/media.repository';
+import { UmbMediaRepository } from '../../repository/media.repository';
 import {
 	UmbModalContext,
 	UMB_MODAL_CONTEXT_TOKEN,
@@ -16,8 +16,6 @@ import type { UmbObserverController } from '@umbraco-cms/backoffice/observable-a
 
 @customElement('umb-input-media-picker')
 export class UmbInputMediaPickerElement extends FormControlMixin(UmbLitElement) {
-	
-
 	/**
 	 * This is a minimum amount of selected items in this input.
 	 * @type {number}
@@ -184,7 +182,7 @@ export class UmbInputMediaPickerElement extends FormControlMixin(UmbLitElement) 
 		`;
 		//TODO: <uui-button-inline-create vertical></uui-button-inline-create>
 	}
-	
+
 	static styles = [
 		UUITextStyles,
 		css`
