+++ src/backoffice/media/media/components/input-media-picker/input-media-picker.stories.ts
