+++ src/backoffice/packages/package-builder/workspace/workspace-package-builder.element.ts
@@ -3,9 +3,10 @@ import { UUIBooleanInputEvent, UUIInputElement, UUIInputEvent } from '@umbraco-u
 import { css, html, nothing } from 'lit';
 import { customElement, property, query, state } from 'lit/decorators.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
-import { UmbInputDocumentPickerElement } from '../../../documents/documents/components/input-document-picker/input-document-picker.element';
-import { UmbInputMediaPickerElement } from '../../../shared/components/input-media-picker/input-media-picker.element';
-import { UmbInputLanguagePickerElement } from '../../../settings/languages/components/input-language-picker/input-language-picker.element';
+// TODO: update to module imports when ready
+import type { UmbInputDocumentPickerElement } from '../../../documents/documents/components/input-document-picker/input-document-picker.element';
+import type { UmbInputMediaPickerElement } from '../../../media/media/components/input-media-picker/input-media-picker.element';
+import type { UmbInputLanguagePickerElement } from '../../../settings/languages/components/input-language-picker/input-language-picker.element';
 import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
 import { PackageDefinitionResponseModel, PackageResource } from '@umbraco-cms/backoffice/backend-api';
 import { tryExecuteAndNotify } from '@umbraco-cms/backoffice/resources';
